requirements.txt
streamlit
requests
bs4
gtts
google-generativeai
python-dotenv
fastapi
uvicorn